glsl, gles2: add shader/frag tests for gles2
This is a set of trivial changes. In the appropriate set of
tests/glslparsertests/shaders/*.frag which are targeted for
glsl_version 1.10 or 1.20, add glsles_version: 1.00 which
identifies the test as appropriate for glsl es 1.0 which lines up
with gles2.

CorrectVersion.V100.frag and CorrectVersion1.V100.frag are new
specific for GLSL ES testing #version 100.