piglit-util-egl: check if EGL_EXT_client_extensions is supported
From the EGL_EXT_client_extensions spec [0]
[0] https://www.khronos.org/registry/egl/extensions/EXT/EGL_EXT_client_extensions.txt
1. How should clients detect if this extension is supported?
   RESOLVED: If an EGL implementation supports this extension, then
   `eglQueryString(EGL_NO_DISPLAY, EGL_EXTENSIONS)` returns
   a well-formed extension string and generates no error.  Otherwise, it
   returns NULL and generates EGL_BAD_DISPLAY.

So, detect the case where EGL_EXT_client_extensions is not supported and
just return false indicating that the requested client extension is also
not supported.

This keeps piglit_is_extension_in_string() from crashing when it tries
to find a needle in a NULL haystack.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>