dispatch: Use dlsym to lookup core symbols for EGL
eglGetProcAddress() only supports extension functions.
Therefore, we must use dlsym() directly on the GL client library to look
up core functions.

The implementation here is a much simplified version of the one in
libepoxy.  Most of the simplification is because piglit dispatch already
knows exactly for which GL API it is looking up a symbol.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>