oes_compressed_paletted_texture: relax glTexImage2D error value expectation
The OES_compressed_paletted_texture spec [0] is very clear that
GL_INVALID_OPERATION should be returned by TexImage2D and friends if
internalformat is one of formats defined in the spec.

[0] https://www.opengl.org/registry/specs/OES/OES_compressed_paletted_texture.txt

It is totally reasonable that some implementations may return that error.
So, allow the test to pass in that case.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>