egl_khr_create_context: add mini egl dispatch to verify-gl-flavor
piglit dispatch is not able to 'reinitialize' in the middle of a test.
Once we've initialized it for a particular dispatch api, it is fixed for
the duration of the test.

This test actually creates contexts of several different client APIs.
Each time, it needs to call some bootstrapping functions (glGetString &
glGetIntegerv) to check that the context is working.  Each time, however,
it needs to call that API's specific bootstrap functions - thus it needs
to do a new "dispatch" everytime.

Furthermore, since these bootstrapping functions are GL "core" functions,
we can't use eglGetProcAddress to fetch them - we have to use dlsym.

This hack is linux specific; a better solution will have to wait for
epoxy.

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>