Update expectations for array-ctor-* tests.
These tests mostly test GLSL functionality which is a no-op in GLSL-ES 1.00
or is supported in GLSL-ES 3.00. For now, update the test expectations to fail
when using  GLSL-ES 1.00. Eventually, we want to split these to GLSL-ES 3.00 tests
or could use some of these to test ES 1.00 is not allowing this functionality.

Also, added comments per test as to why this is expected to fail or pass.